update generator printing, babylon [skip ci]
