+++ packages/babel-types/package.json
@@ -14,6 +14,6 @@
   },
   "devDependencies": {
     "babel-generator": "7.0.0-beta.1",
-    "babylon": "^7.0.0-beta.22"
+    "babylon": "^7.0.0-beta.25"
   }
 }
